Rename Pattern to SmithyPattern
SmithyPattern is a lot less annoying to use since it doesn't conflict
with the Java regex Pattern.
